backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.1",
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.6",
    "langchain-openai>=1.1.7",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.40.0",
]
